LPS-77835 Move JspServlet to internal package.
