+++ modules/apps/static/portal-osgi-web/source-formatter.properties
@@ -6,4 +6,4 @@
 ##
 
     proxy.excludes=\
-        modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/main/java/com/liferay/portal/osgi/web/servlet/jsp/compiler/JspServlet.java
+        modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-jsp-compiler/src/main/java/com/liferay/portal/osgi/web/servlet/jsp/compiler/internal/JspServlet.java
